Silence new VS Windows warnings
src\game\client\components\menus_settings.cpp(2516,26): warning C5055: operator '/': deprecated between enumerations and floating-point types

+++ src/game/client/components/menus_settings.cpp
@@ -2513,11 +2513,11 @@ void CMenus::RenderSettingsAppearance(CUIRect MainView)
 
 	MainView.HSplitTop(20, &TabBar, &MainView);
 	float TabsW = TabBar.w;
-	TabBar.VSplitLeft(TabsW / NUMBER_OF_APPEARANCE_TABS, &Page1Tab, &Page2Tab);
-	Page2Tab.VSplitLeft(TabsW / NUMBER_OF_APPEARANCE_TABS, &Page2Tab, &Page3Tab);
-	Page3Tab.VSplitLeft(TabsW / NUMBER_OF_APPEARANCE_TABS, &Page3Tab, &Page4Tab);
-	Page4Tab.VSplitLeft(TabsW / NUMBER_OF_APPEARANCE_TABS, &Page4Tab, &Page5Tab);
-	Page5Tab.VSplitLeft(TabsW / NUMBER_OF_APPEARANCE_TABS, &Page5Tab, &Page6Tab);
+	TabBar.VSplitLeft(TabsW / static_cast<float>(NUMBER_OF_APPEARANCE_TABS), &Page1Tab, &Page2Tab);
+	Page2Tab.VSplitLeft(TabsW / static_cast<float>(NUMBER_OF_APPEARANCE_TABS), &Page2Tab, &Page3Tab);
+	Page3Tab.VSplitLeft(TabsW / static_cast<float>(NUMBER_OF_APPEARANCE_TABS), &Page3Tab, &Page4Tab);
+	Page4Tab.VSplitLeft(TabsW / static_cast<float>(NUMBER_OF_APPEARANCE_TABS), &Page4Tab, &Page5Tab);
+	Page5Tab.VSplitLeft(TabsW / static_cast<float>(NUMBER_OF_APPEARANCE_TABS), &Page5Tab, &Page6Tab);
 
 	static CButtonContainer s_aPageTabs[NUMBER_OF_APPEARANCE_TABS] = {};
